Filebeat Patch for Datadog Integration
